hyraigne/site/mangadex: handle decimal chapters
Some chapter ID in MangaDex have a fractional part, like 3.5 or 99.1,
and those weren't formatted correctly because the specified width
applies to the whole number, not only the integral part.

In the end, we got "3.5" instead of "003.5" and we don't have 3.5
between 3 ("003") and 4 ("004") on the filesystem.

This commit introduces a function to format the ID with a fractional
part as expected.