bibe: add MangaDex support
This requires to expose two new filter options:
- language
- group

For now those are only useful for MangaDex, to filter by language and
to decide which version of a chapter to select when there are multiple
versions by different teams.

Also add a log to warn the user when the filter returned no chapters
(could happen with language filtering).